M3-412 Feat : 동시성 문제 해결
